requirements.txt
vanna[gemini,sqlite]>=2.0.0,<2.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
plotly>=5.24.0
pandas>=2.2.0
python-dotenv>=1.0.0
google-genai>=1.0.0
